[Fix] confusion between using unique ptr for Field
